enable running tests on real cluster
This change refactors pseudo_hdfs4 to introduce a new class RealCluster that shares a parent class, ClusterBase, with PseudoHdfs4. Based on an environment variable, HUE_TEST_ON_CLUSTER, RealCluster is used instead of PseudoHdfs4. In that case, instead of starting pseudo cluster processes and mocking things, the cluster pointed to by the effective Hue config is used. That is, for example, self.fs points to a previously set up WebHdfs endpoint instead of a local one that's set up for the test.

Note that this should not affect existing functionality or tests in any way. Also, further changes will be made to make all tests pass on a real cluster, such as conditionally replacing mock Apis like "shared oozie server".